Base page objects are now proper page objects and not just base classes This required moving the `pageobjects` decorator out of `__init__.py` Also improved how we're finding base classes when a generic page object is requested.
